--- src/components/Field/FormFieldChipSingleSelect/index.tsx
@@ -1,3 +1,0 @@
-export { default } from "./FormFieldChipSingleSelect";
-export * from "./FormFieldChipSingleSelect";
-export * from "./FormFieldChipSingleSelectTypes";
