+++ src/components/Field/FormFieldChips/FormFieldChips.stories.mdx
@@ -1,11 +1,11 @@
 import { Meta } from '@storybook/addon-docs/blocks';
 
-<Meta title='FormFields/FormFieldChipSingleSelect/Readme' />
+<Meta title='FormFields/FormFieldChips/Readme' />
 
 # Chip Single Select
 This field implements the [**FieldDef**](#generic-field-props-fielddef) interface.
 
-- The `FormFieldChipSingleSelect` component is built over a wrapper for [MUI Autocomplete](https://mui.com/material-ui/react-autocomplete/#main-content) but with SimpleView brand colors.
+- The `FormFieldChips` component is built over a wrapper for [MUI Autocomplete](https://mui.com/material-ui/react-autocomplete/#main-content) but with SimpleView brand colors.
 - [**Playground**](/?path=/story/formfields-formfieldchipsingleselect--playground)
 - Data: `string` - Value of the selected option.
 - inputSettings:
