+++ src/components/Field/FormFieldChips/index.tsx
@@ -1,0 +1,3 @@
+export { default } from "./FormFieldChips";
+export * from "./FormFieldChips";
+export * from "./FormFieldChipsTypes";
